更新

+++ serve/app/Http/Controllers/Apps/Sms/SmsController.php
@@ -19,7 +19,7 @@ class SmsController extends Controller
         $datas = new SysSmsTemplate();
         $datas = $datas->get()->groupBy('template_type');
         $sms_templates = array();
-        foreach($datas as $key=>$data){
+        foreach ($datas as $key => $data) {
             $sms_templates[$key] = SmsResource::collection($data);
         }
         return $this->jsonSuccessResponse($sms_templates);
@@ -30,9 +30,9 @@ class SmsController extends Controller
         $shop = $request->get('ori_shop');
         $sms_template = SysSmsTemplate::findOrFail($request->route()->parameter('sms'));
         $shop_sms_template = $shop->sms_templates()->find($sms_template['id']);
-        if($shop_sms_template){
+        if ($shop_sms_template) {
             $shop->sms_templates()->detach($sms_template['id']);
-        }else{
+        } else {
             $shop->sms_templates()->attach($sms_template['id']);
         }
         $shop->refresh();
@@ -49,8 +49,8 @@ class SmsController extends Controller
     public function sign_store(Request $request)
     {
         $shop = $request->get('ori_shop');
-        $validator = Validator::make($request->all(),[
-            "sign_name"=>"required"
+        $validator = Validator::make($request->all(), [
+            "sign_name" => "required"
         ]);
         if ($validator->fails()) {
             return $this->jsonSuccessResponse($validator->errors()->first());
@@ -66,6 +66,7 @@ class SmsController extends Controller
     {
         $shop = $request->get('ori_shop');
         $sign = $shop->signs()->findOrFail($request->route()->parameter('sign'));
+        if ($sign['status'] == ShopSmsSign::SIGN_STATUS_SUCCESS) return $this->jsonErrorResponse(422, "审核通过的签名不能删除");
         $sign->delete();
         return $this->jsonSuccessResponse();
     }
@@ -74,15 +75,19 @@ class SmsController extends Controller
     {
         $shop = $request->get('ori_shop');
         $sign = $shop->signs()->findOrFail($request->route()->parameter('sign'));
-        if($sign['status'] != ShopSmsSign::SIGN_STATUS_SUCCESS) return $this->jsonErrorResponse(422,"该签名未通过审核，不能启用");
-        $validator = Validator::make($request->all(),[
-            "active"=>"required|boolean"
+        if ($sign['status'] != ShopSmsSign::SIGN_STATUS_SUCCESS) return $this->jsonErrorResponse(422, "该签名未通过审核，不能启用");
+        $validator = Validator::make($request->all(), [
+            "active" => "required|boolean"
         ]);
         if ($validator->fails()) {
             return $this->jsonSuccessResponse($validator->errors()->first());
         } else {
             $data = $validator->validate();
-            if(isset($data['active']) && $data('active')) $shop->signs()->update('active',false);
+            if (isset($data['active']) && $data('active')) {
+                if ($sign['status'] == ShopSmsSign::SIGN_STATUS_FAILED)
+                    return $this->jsonErrorResponse(422, "未通过审核的签名不能启用");
+                $shop->signs()->update('active', false);
+            }
             $sign = $sign->update($data);
         }
         return $this->jsonSuccessResponse(new SmsSignResource($sign));
@@ -93,11 +98,9 @@ class SmsController extends Controller
         $shop = $request->get('ori_shop');
         $logs = $shop->sms_logs();
 
-        $logs = $logs->orderBy('created_at','desc')->paginate($request->get('pageSize'));
+        $logs = $logs->orderBy('created_at', 'desc')->paginate($request->get('pageSize'));
         return $this->jsonSuccessResponse(new SmsLogListCollection($logs));
     }
-
-
 
 
 }
